+++ contracts/handlers/balancer/HBalancer.sol
@@ -31,7 +31,7 @@ contract HBalancer is HandlerBase {
         uint256 beforePoolAmount = IERC20(pool).balanceOf(address(this));
 
         // Execute "joinswapExternAmountIn" by using DSProxy
-        tokenAmountIn = _getProxyBalance(tokenIn, tokenAmountIn);
+        tokenAmountIn = _getBalance(tokenIn, tokenAmountIn);
         IERC20(tokenIn).safeApprove(address(proxy), tokenAmountIn);
         try
             proxy.execute(
@@ -75,7 +75,7 @@ contract HBalancer is HandlerBase {
 
         // Approve all erc20 token to Proxy
         for (uint256 i = 0; i < tokens.length; i++) {
-            maxAmountsIn[i] = _getProxyBalance(tokens[i], maxAmountsIn[i]);
+            maxAmountsIn[i] = _getBalance(tokens[i], maxAmountsIn[i]);
             IERC20(tokens[i]).safeApprove(address(proxy), maxAmountsIn[i]);
             amountsIn[i] = IERC20(tokens[i]).balanceOf(address(this));
         }
@@ -121,7 +121,7 @@ contract HBalancer is HandlerBase {
         IBPool bPool = IBPool(pool);
 
         // Call exitswapPoolAmountIn function of balancer pool
-        poolAmountIn = _getProxyBalance(address(bPool), poolAmountIn);
+        poolAmountIn = _getBalance(address(bPool), poolAmountIn);
         try
             bPool.exitswapPoolAmountIn(tokenOut, poolAmountIn, minAmountOut)
         returns (uint256 amount) {
@@ -155,7 +155,7 @@ contract HBalancer is HandlerBase {
         }
 
         // Call exitPool function of balancer pool
-        poolAmountIn = _getProxyBalance(address(bPool), poolAmountIn);
+        poolAmountIn = _getBalance(address(bPool), poolAmountIn);
         try bPool.exitPool(poolAmountIn, minAmountsOut) {} catch Error(
             string memory reason
         ) {