+++ contracts/handlers/compound/HCToken.sol
@@ -25,7 +25,7 @@ contract HCToken is HandlerBase {
 
         address token = _getToken(cToken);
         // if amount == uint256(-1) return balance of Proxy
-        mintAmount = _getProxyBalance(token, mintAmount);
+        mintAmount = _getBalance(token, mintAmount);
         IERC20(token).safeApprove(cToken, mintAmount);
         try compound.mint(mintAmount) returns (uint256 errorCode) {
             if (errorCode != 0)
@@ -59,7 +59,7 @@ contract HCToken is HandlerBase {
 
         ICToken compound = ICToken(cToken);
         // if amount == uint256(-1) return balance of Proxy
-        redeemTokens = _getProxyBalance(cToken, redeemTokens);
+        redeemTokens = _getBalance(cToken, redeemTokens);
         IERC20(cToken).safeApprove(cToken, redeemTokens);
         try compound.redeem(redeemTokens) returns (uint256 errorCode) {
             if (errorCode != 0)