+++ contracts/handlers/curve/HCurve.sol
@@ -30,7 +30,7 @@ contract HCurve is HandlerBase {
         ICurveHandler curveHandler = ICurveHandler(handler);
         uint256 beforeTokenJBalance = IERC20(tokenJ).balanceOf(address(this));
         // if amount == uint256(-1) return balance of Proxy
-        dx = _getProxyBalance(tokenI, dx);
+        dx = _getBalance(tokenI, dx);
         IERC20(tokenI).safeApprove(address(curveHandler), dx);
         try curveHandler.exchange(i, j, dx, minDy) {} catch Error(
             string memory reason
@@ -59,7 +59,7 @@ contract HCurve is HandlerBase {
         ICurveHandler curveHandler = ICurveHandler(handler);
         uint256 beforeTokenJBalance = IERC20(tokenJ).balanceOf(address(this));
         // if amount == uint256(-1) return balance of Proxy
-        dx = _getProxyBalance(tokenI, dx);
+        dx = _getBalance(tokenI, dx);
         IERC20(tokenI).safeApprove(address(curveHandler), dx);
         try curveHandler.exchange_underlying(i, j, dx, minDy) {} catch Error(
             string memory reason
@@ -87,7 +87,7 @@ contract HCurve is HandlerBase {
         IOneSplit oneSplit = IOneSplit(ONE_SPLIT);
         uint256 beforeToTokenBalance = IERC20(toToken).balanceOf(address(this));
         // if amount == uint256(-1) return balance of Proxy
-        amount = _getProxyBalance(fromToken, amount);
+        amount = _getBalance(fromToken, amount);
         IERC20(fromToken).safeApprove(address(oneSplit), amount);
         try
             oneSplit.swap(
@@ -114,7 +114,7 @@ contract HCurve is HandlerBase {
     function addLiquidity(
         address handler,
         address pool,
-        address[] memory tokens,
+        address[] calldata tokens,
         uint256[] memory amounts,
         uint256 minMintAmount
     ) external payable returns (uint256) {
@@ -125,7 +125,7 @@ contract HCurve is HandlerBase {
         for (uint256 i = 0; i < amounts.length; i++) {
             if (amounts[i] == 0) continue;
             // if amount == uint256(-1) return balance of Proxy
-            amounts[i] = _getProxyBalance(tokens[i], amounts[i]);
+            amounts[i] = _getBalance(tokens[i], amounts[i]);
             IERC20(tokens[i]).safeApprove(address(curveHandler), amounts[i]);
         }