+++ contracts/handlers/maker/HMaker.sol
@@ -44,7 +44,7 @@ contract HMaker is HandlerBase {
     ) external payable returns (uint256 cdp) {
         IDSProxy proxy = IDSProxy(_getProxy(address(this)));
         // if amount == uint256(-1) return balance of Proxy
-        value = _getProxyBalance(address(0), value);
+        value = _getBalance(address(0), value);
 
         try
             proxy.execute.value(value)(
@@ -85,7 +85,7 @@ contract HMaker is HandlerBase {
         address token = IMakerGemJoin(gemJoin).gem();
 
         // if amount == uint256(-1) return balance of Proxy
-        wadC = _getProxyBalance(token, wadC);
+        wadC = _getBalance(token, wadC);
 
         IERC20(token).safeApprove(address(proxy), wadC);
         try
@@ -127,7 +127,7 @@ contract HMaker is HandlerBase {
         IDSProxy proxy = IDSProxy(_getProxy(address(this)));
         address owner = _getProxy(_getSender());
         // if amount == uint256(-1) return balance of Proxy
-        value = _getProxyBalance(address(0), value);
+        value = _getBalance(address(0), value);
 
         try
             proxy.execute.value(value)(
@@ -157,7 +157,7 @@ contract HMaker is HandlerBase {
         address owner = _getProxy(_getSender());
         address token = IMakerGemJoin(gemJoin).gem();
         // if amount == uint256(-1) return balance of Proxy
-        wad = _getProxyBalance(token, wad);
+        wad = _getBalance(token, wad);
         IERC20(token).safeApprove(address(proxy), wad);
         try
             proxy.execute(