+++ contracts/handlers/mooniswap/HMooniswap.sol
@@ -19,9 +19,9 @@ contract HMooniswap is HandlerBase {
     }
 
     function deposit(
-        address[2] memory tokens,
+        address[2] calldata tokens,
         uint256[] memory amounts,
-        uint256[] memory minAmounts
+        uint256[] calldata minAmounts
     ) external payable returns (uint256 fairSupply) {
         if (tokens[0] > tokens[1]) _revertMsg("deposit", "wrong tokens order");
         if (tokens[0] == tokens[1]) _revertMsg("deposit", "same tokens");
@@ -32,8 +32,8 @@ contract HMooniswap is HandlerBase {
         IMooniswap mooniswap = IMooniswap(factory.pools(tokens[0], tokens[1]));
 
         // Approve token
-        amounts[0] = _getProxyBalance(tokens[0], amounts[0]);
-        amounts[1] = _getProxyBalance(tokens[1], amounts[1]);
+        amounts[0] = _getBalance(tokens[0], amounts[0]);
+        amounts[1] = _getBalance(tokens[1], amounts[1]);
         uint256 value = 0;
         if (tokens[0] == address(0)) {
             value = amounts[0];
@@ -83,7 +83,7 @@ contract HMooniswap is HandlerBase {
         }
 
         // Remove liquidity
-        amount = _getProxyBalance(pool, amount);
+        amount = _getBalance(pool, amount);
         try mooniswap.withdraw(amount, minReturns) {} catch Error(
             string memory reason
         ) {